refactor: move generator scripts to lib/ folder
