+++ package.json
@@ -11,9 +11,9 @@
   "scripts": {
     "commitlint": "commitlint --from=$(git cherry origin/main | head -n 1 | cut -c 3-)",
     "gen": "npm run gen-docs && npm run gen-enums && npm run gen-types && npm run gen-cjs",
-    "gen-docs": "node src/gen-docs.js",
-    "gen-enums": "node src/gen-enums.js",
-    "gen-types": "node src/gen-types.js",
+    "gen-docs": "node src/lib/gen-docs.js",
+    "gen-enums": "node src/lib/gen-enums.js",
+    "gen-types": "node src/lib/gen-types.js",
     "gen-cjs": "npx rollup ./src/index.js --file ./src/index.cjs --format cjs && npx rollup ./src/enums.js --file ./src/enums.cjs --format cjs",
     "lint": "npx eslint .",
     "format": "prettier --write .",
